requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
pandas
fastapi
flask
ipykernel
numpy
python-dotenv
python-multipart
requests
setuptools
uvicorn
streamlit